Refactor resultsprocessing package
