+++ QueryBuilder_Cassandra/src/main/java/net/sf/esfinge/querybuilder/cassandra/querybuilding/resultsprocessing/ordering/ChainComparatorFactory.java
@@ -1,4 +1,4 @@
-package net.sf.esfinge.querybuilder.cassandra.querybuilding.ordering;
+package net.sf.esfinge.querybuilder.cassandra.querybuilding.resultsprocessing.ordering;
 
 import net.sf.esfinge.querybuilder.cassandra.reflection.CassandraReflectionUtils;
